+++ NEWS.md
@@ -1,5 +1,6 @@
 # idaifieldR 0.2.3 _2023-02-18_
 * add language list lookup preparation
+* add language management for input fields
 * new demodata
 
 # idaifieldR 0.2.2 _2023-02-18_